feat: Initial commit
This commit introduces a minimal NixOS configuration for fresh NixOS hosts.

The configuration is designed to be as simple as possible, with good defaults for all systems. It creates a user, enables SSH with root access, and uses public keys from a `public-keys.nix` file.

The flake exposes several configurations for different architectures and bootloaders:

- `x86_64-grub`: for x86_64 systems using GRUB with legacy BIOS.
- `x86_64-systemd`: for x86_64 systems using systemd-boot with UEFI.
- `aarch64-systemd`: for aarch64 systems using systemd-boot with UEFI.
- `aarch64-uboot`: for aarch64 systems using U-Boot.

The user is required to generate their own hardware configuration using `nixos-generate-config`.